+++ lib/project_types/theme/commands/push.rb
@@ -4,13 +4,21 @@ require "shopify_cli/theme/development_theme"
 require "shopify_cli/theme/ignore_filter"
 require "shopify_cli/theme/include_filter"
 require "shopify_cli/theme/syncer"
+require "project_types/theme/commands/common/root_helper"
+require "project_types/theme/conversions/include_glob"
+require "project_types/theme/conversions/ignore_glob"
 
 module Theme
   class Command
     class Push < ShopifyCLI::Command::SubCommand
+      include Common::RootHelper
+
       recommend_default_ruby_range
 
       options do |parser, flags|
+        Conversions::IncludeGlob.register(parser)
+        Conversions::IgnoreGlob.register(parser)
+
         parser.on("-n", "--nodelete") { flags[:nodelete] = true }
         parser.on("-i", "--themeid=ID") { |theme_id| flags[:theme_id] = theme_id }
         parser.on("-t", "--theme=NAME_OR_ID") { |theme| flags[:theme] = theme }
@@ -20,18 +28,18 @@ module Theme
         parser.on("-j", "--json") { flags[:json] = true }
         parser.on("-a", "--allow-live") { flags[:allow_live] = true }
         parser.on("-p", "--publish") { flags[:publish] = true }
-        parser.on("-o", "--only=PATTERN") do |pattern|
+        parser.on("-o", "--only=PATTERN", Conversions::IncludeGlob) do |pattern|
           flags[:includes] ||= []
-          flags[:includes] << pattern
+          flags[:includes] += pattern
         end
-        parser.on("-x", "--ignore=PATTERN") do |pattern|
+        parser.on("-x", "--ignore=PATTERN", Conversions::IgnoreGlob) do |pattern|
           flags[:ignores] ||= []
-          flags[:ignores] << pattern
+          flags[:ignores] += pattern
         end
       end
 
-      def call(args, _name)
-        root = args.first || "."
+      def call(_args, name)
+        root = root_value(options, name)
         delete = !options.flags[:nodelete]
         theme = find_theme(root, **options.flags)
         return if theme.nil?
@@ -42,7 +50,7 @@ module Theme
           return unless CLI::UI::Prompt.confirm(question)
         end
 
-        include_filter = ShopifyCLI::Theme::IncludeFilter.new(options.flags[:includes])
+        include_filter = ShopifyCLI::Theme::IncludeFilter.new(root, options.flags[:includes])
         ignore_filter = ShopifyCLI::Theme::IgnoreFilter.from_path(root)
         ignore_filter.add_patterns(options.flags[:ignores]) if options.flags[:ignores]